+++ src/main/java/cech12/extendedmushrooms/block/InfestedFlowerBlock.java
@@ -6,6 +6,7 @@ import net.minecraft.block.Blocks;
 import net.minecraft.block.FlowerBlock;
 import net.minecraft.particles.ParticleTypes;
 import net.minecraft.potion.Effect;
+import net.minecraft.util.Direction;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.vector.Vector3d;
 import net.minecraft.world.IBlockReader;
@@ -15,8 +16,6 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
 import java.util.Random;
-
-import net.minecraft.block.AbstractBlock.Properties;
 
 public class InfestedFlowerBlock extends FlowerBlock {
 
@@ -50,5 +49,15 @@ public class InfestedFlowerBlock extends FlowerBlock {
                     blockPos.getY() + offset.y + (random.nextDouble() * 0.2),
                     blockPos.getZ() + offset.z + (random.nextDouble() * 0.375), 0.0D, 0.0D, 0.0D);
         }
+    }
+
+    @Override
+    public int getFlammability(BlockState state, IBlockReader world, BlockPos pos, Direction face) {
+        return 100;
+    }
+
+    @Override
+    public int getFireSpreadSpeed(BlockState state, IBlockReader world, BlockPos pos, Direction face) {
+        return 60;
     }
 }