+++ src/main/java/cech12/extendedmushrooms/item/MushroomWoodType.java
@@ -57,7 +57,7 @@ public enum MushroomWoodType implements IStringSerializable {
         this.planksBlock = planksBlock;
         this.boatItem = boatItem;
         this.lightValue = lightValue;
-        this.woodType = WoodType.create(new ResourceLocation(ExtendedMushrooms.MOD_ID, name).toString());
+        this.woodType = WoodType.register(WoodType.create(new ResourceLocation(ExtendedMushrooms.MOD_ID, name).toString()));
     }
 
     public int getId() {