+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.16.5-1.7.0.3] - 2022-07-26
+### Fixed
+- Mushroom wood signs were not rendered correctly
+- Slime Fungus Baashrooms dropped the wrong mushroom cap block
+- Honey Fungus & Slime Fungus blocks were not compostable
+- Infested Flower was not flammable
+- Honey Fungus chests & signs were not registered correctly
+
 ## [1.16.5-1.7.0.2] - 2022-06-30
 ### Fixed
 - Server crashed when placing an item into Fairy Ring.